Tir avec clic gauche et utilisation d'ObservableList+ListChangeListener

+++ Terraria/src/application/modele/item/Projectile.java
@@ -3,12 +3,14 @@ package application.modele.item;
 import application.modele.Environnement;
 import application.modele.personnage.Perso;
 import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleIntegerProperty;
 
 public class Projectile extends Arme{
 
-	private DoubleProperty x;
-	private DoubleProperty y;
+	private IntegerProperty x;
+	private IntegerProperty y;
 	private double trajectoirX;
 	private double trajectoirY;
 	private Perso p;
@@ -16,18 +18,18 @@ public class Projectile extends Arme{
 	public Projectile(int id,  Environnement env, Perso p, int xDest, int yDest) {
 		super(id, 25, env);
 		this.p=p;
-		this.x= new SimpleDoubleProperty(this.p.getX()/*128*/);
-		this.y= new SimpleDoubleProperty(this.p.getY()/*128*/);
+		this.x= new SimpleIntegerProperty(this.p.getX()/*128*/);
+		this.y= new SimpleIntegerProperty(this.p.getY()/*128*/);
 		
-		this.trajectoirX=((xDest-x.get())/p.getEnv().getColonne() );
+		this.trajectoirX=((xDest-x.get())/p.getEnv().getColonne());
 		this.trajectoirY=((yDest-y.get())/p.getEnv().getLigne());
 	}
 	
-	public Double getX() {
+	public int getX() {
 		return this.x.getValue();
 	}
 	
-	public DoubleProperty getXProperty() {
+	public IntegerProperty getXProperty() {
 		return x;
 	}
 	
@@ -35,11 +37,11 @@ public class Projectile extends Arme{
 		this.x.setValue(this.getX()+xPlus);
 	}
 	
-	public Double getY() {
+	public int getY() {
 		return this.y.getValue();
 	}
 	
-	public DoubleProperty getYProperty() {
+	public IntegerProperty getYProperty() {
 		return y;
 	}
 	
@@ -51,8 +53,8 @@ public class Projectile extends Arme{
 		System.out.println(trajectoirX);
 		System.out.println(trajectoirY);
 
-		x.set(x.get()+trajectoirX);
-		y.set(y.get()+trajectoirY);
+		x.set((int) (x.get()+trajectoirX));
+		y.set((int) (y.get()+trajectoirY));
 	}
 	
 }
